@@ -1,32 +1,42 @@
-const Author = require('../models/author.model');
+const Author = require("../models/author");
 
-module.exports.createAuthor = (request, response) => {
-    const { name } = request.body;
-    Author.create({name})
-        .then(res => response.json(res))
-        .catch(err => response.status(400).json(err))
+////////// req = request | res = response \\\\\\\\\\
+
+/////  CREATE \\\\\
+
+module.exports.createAuthor = (req, res) => {
+    const { name } = req.body;
+    Author.create({ name })
+        .then(author => res.json(author))
+        .catch(err => res.status(400).json(err));
 }
 
-module.exports.getAllAuthors = (request, response) => {
-    Author.find({})
-        .then(res => response.json(res))
-        .catch(err => response.json(err))
+///// RETRIEVE \\\\\
+
+module.exports.findAllAuthors = (req, res) => {
+    Author.find()
+        .then(allAuthors => res.json(allAuthors))
+        .catch(err => res.status(400).json({ message: "Error! no author for you!", error: err }));
 }
 
-module.exports.getSingleAuthor = (request, response) => {
-    Author.findOne({_id:request.params.id})
-        .then(res => response.json(res))
-        .catch(err => response.json(err))
+module.exports.findOneAuthor = (req, res) => {
+    Author.findById(req.params.id)
+        .then(author => res.json(author))
+        .catch(err => res.status(400).json({ message: "Error! no author for you!", error: err }));
 }
 
-module.exports.updateAuthor = (request, response) => {
-    Author.findOneAndUpdate({_id: request.params.id}, request.body, {new:true})
-        .then(res => response.json(res))
-        .catch(err => response.status(400).json(err))
+///// UPDATE \\\\\
+
+module.exports.updateAuthor = (req, res) => {
+    Author.findOneAndUpdate({ _id: req.params.id }, req.body, { new: true, runValidators: true })
+        .then(updatedAuthor => res.json(updatedAuthor))
+        .catch(err => res.status(400).json(err));
 }
 
-module.exports.deleteAuthor = (request, response) => {
-    Author.deleteOne({ _id: request.params.id })
-        .then(deleteConfirmation => response.json(deleteConfirmation))
-        .catch(err => response.json(err))
+///// DELETE \\\\\
+
+module.exports.deleteAuthor = (req, res) => {
+    Author.findByIdAndDelete(req.params.id)
+        .then(delAuthor => res.json(delAuthor))
+        .catch(err => res.status(400).json({ message: "Error! no author for you!", error: err }));
 }
