@@ -1,7 +1,8 @@
-const mongoose = require('mongoose');
-mongoose.connect("mongodb://localhost/products", {
-    useNewUrlParser: true,
-    useUnifiedTopology: true,
-})
-    .then(() => console.log("Established a connection to the database"))
-    .catch(err => console.log("Something went wrong when connecting to the database", err));
+const mongoose = require("mongoose")
+
+const dbName = process.env.DB
+
+
+mongoose.connect(`mongodb://127.0.0.1:27017/${dbName}`)
+    .then(() => { console.log(`Established a connection to the database ${dbName}`) })
+    .catch(err => console.log("Something Went wrong when connecting to the DB ", err))
